Move HelmChart handling to separate reconciler
This moves the HelmChart template handling to a separate reconciler,
with predicates detecing relevant changes. The idea is that this would
both facilitate working _without_ chart templates but with references
in the future, and to reduce cognitive load while working with
reconciler logic.

The predicate uses cmp.Equal to inspect the Chart template objects of
the old and new HelmRelease object in the update vent.

The reconciler uses cmp.Diff with a custom SimpleReporter, to reduce
the output to a minimum. Example:

```
//  Path:
//  -old
//  +new
//
//  Path.Nested:
//  -removed
//
//  Path.Other:
//  +added
```

The implementation on the release reconciler's end is a rough sketch,
but in working shape. The foresight is that much of the reconciler will
change when the release logic will be adjusted to work with the earlier
introduced storage observer.

Signed-off-by: Hidde Beydals <hello@hidde.co>